Remove `include/grpc/impl/codegen/log.h`
