+++ test/cpp/end2end/message_allocator_end2end_test.cc
@@ -28,7 +28,7 @@
 #include <google/protobuf/arena.h>
 #include <gtest/gtest.h>
 
-#include <grpc/impl/codegen/log.h>
+#include <grpc/support/log.h>
 #include <grpcpp/channel.h>
 #include <grpcpp/client_context.h>
 #include <grpcpp/create_channel.h>
